css/css-scroll-snap-2/snapchanged/snapchanged-with-proximity-strictness.tentative.html

<!DOCTYPE html>
<html>
<head>
<title>
  <title> CSS Scroll Snap 2 Test: snapchanged events on proximity strictness container</title>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#snap-events"/>
</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/dom/events/scrolling/scroll_support.js"></script>
<style>
  div {
    position: absolute;
    margin: 0;
  }

  #scroller {
    height: 600px;
    width: 600px;
    overflow: hidden;
    scroll-snap-type: y proximity;
  }

  .snap {
    width: 300px;
    height: 300px;
    background-color: green;
    scroll-snap-align: start;
  }

  .area {
    width: 2000px;
    height: 2000px;
  }
  </style>
</head>
<body>
<div id="scroller">
  <div class="area"></div>
  <div id="target" class="snap" style="top: 0px;"></div>
</div>

<script>
  let resolve_func = null;

  let snap_targets_seen = [];
  function store_snaptargets(evt) {
    for (const element of evt.snapTargets) {
      snap_targets_seen.push(element.id);
    }
    resolve_func();
  }
  promise_test(async (test) => {
    await waitForCompositorCommit();
    // The initial snap position is at (0, 0).
    assert_equals(scroller.scrollTop, 0);
    assert_equals(scroller.scrollLeft, 0);

    scroller.addEventListener("snapchanged", store_snaptargets);
    let snapchanged_promise = new Promise(resolve => resolve_func = resolve);

    // Scroll to a position where it's outside of the scroll-snap proximity
    // threshold, so that it won't trigger snapping.
    scroller.scrollTo(0, 250);

    // snapchanged should fire as we've moved from within the proximity range
    // to outside the proximity range and are no longer snapped.
    await snapchanged_promise;
    assert_equals(scroller.scrollTop, 250);
    assert_equals(snap_targets_seen.length, 0, "scroller is not snapped to any " +
        " element");

    snapchanged_promise = new Promise(resolve => resolve_func = resolve);
    // Scroll to a position within the scroll-snap proximity
    // threshold, so that it triggers snapping.
    scroller.scrollTo(0, 190);

    // snapchanged should fire as we've moved from outside the proximity range
    // to inside the proximity range and are once again snapped.
    await snapchanged_promise;

    assert_equals(scroller.scrollTop, 0);
    assert_equals(snap_targets_seen.length, 1, "scroller is snapped to target");
    assert_equals(snap_targets_seen[0], target.id, "scroller is snapped to " +
        "target");
  }, "Snapchanged fires when scrolling outside proximity range.");
  </script>
</body>
</html>
